tool: add support for passing CNodes to PDs
The CNodes are specified in the system description file and attached
to the PD's CSpace. This allows a PD given Untypeds to create and
manage its own kernel objects outside of Microkit.

Normally PDs are provided with a single rather small CNode that has
many guard bits. PDs do not have capabilities to their own root CNode
so they cannot store new capabilities there. We also cannot insert a
CNode since it would not have enough bits of address its CSlots. The
trick here is to create a new root CNode for PDs, inserting the
previous root CNode as slot zero, thus preserving all indexes. The new
CNodes defined in the system description file are registered in the
following CSlots of the new root CNode, with the given guard size.

Signed-off-by: Mathieu Mirmont <mat@neutrality.ch>